update file names within pages directory
