+++ src/components/learning-center/learning-searchbar.tsx
@@ -10,7 +10,7 @@ import {
 } from "react-instantsearch-dom";
 import { Link } from "gatsby";
 import { SearchBoxExposed } from "react-instantsearch-core";
-import { Locale } from "../../pages";
+import { Locale } from "../../pages/index.en";
 import { I18n } from "@lingui/react";
 import { t, Trans } from "@lingui/macro";
