+++ src/components/learning-center/thank-you-banner.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { documentToReactComponents } from "@contentful/rich-text-react-renderer";
-import { ContentfulContent } from "../../pages";
+import { ContentfulContent } from "../../pages/index.en";
 
 export const ThankYouBanner = (props: ContentfulContent) => (
   <section className="hero is-small has-background-light is-paddingless">
