+++ src/pages/about/team.en.tsx
@@ -8,7 +8,7 @@ import "../../styles/team.scss";
 
 import Layout from "../../components/layout";
 import ReadMore from "../../components/read-more";
-import { ContentfulContent } from "..";
+import { ContentfulContent } from "../index.en";
 
 export const TeamPageScaffolding = (props: ContentfulContent) => (
   <Layout metadata={props.content.metadata} locale={props.locale}>
