+++ src/pages/about/team.es.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { StaticQuery, graphql } from "gatsby";
-import { TeamPageScaffolding } from "./team";
+import { TeamPageScaffolding } from "./team.en";
 
 const TeamPage = () => (
   <StaticQuery
