+++ src/pages/changes.es.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { StaticQuery, graphql } from "gatsby";
-import { ChangelogPageScaffolding } from "./changes";
+import { ChangelogPageScaffolding } from "./changes.en";
 
 const SpanishChangelogPage: React.FC<{}> = () => (
   <StaticQuery
