+++ src/pages/privacy-policy-norent.tsx
@@ -3,7 +3,7 @@ import { StaticQuery, graphql } from "gatsby";
 
 import Layout from "../components/layout";
 import { documentToReactComponents } from "@contentful/rich-text-react-renderer";
-import { ContentfulContent } from ".";
+import { ContentfulContent } from "./index.en";
 
 export const NorentPrivacyPolicyPageScaffolding = (
   props: ContentfulContent
